refactor(tests): move v3 meter tests to import file
